Retry work
