+++ src/main/java/daco/DacoException.java
@@ -1,8 +1,8 @@
-package Daco.Ui;
+package daco;
 
 public class DacoException extends Exception {
-    public enum ErrorType {DOES_NOT_EXIST, INVALID_COMMANDMARK, UNKNOWN_COMMAND, INVALID_NUMBER, EMPTY_TASK, EMPTY_DATE}
-    private final ErrorType type;
+    public enum ErrorType {DOES_NOT_EXIST, INVALID_COMMANDMARK, UNKNOWN_COMMAND, INVALID_NUMBER, EMPTY_TASK, EMPTY_DATE, INVALID_FORMAT_DELETE}
+    protected final ErrorType type;
 
     public DacoException(ErrorType type) {
         super("");
