+++ src/main/java/daco/Event.java
@@ -1,6 +1,5 @@
-package Daco.Task;
-import Daco.Ui.DateAndTime;
-import Daco.Ui.DacoException;
+package daco;
+
 public class Event extends Task {
     protected DateAndTime duration;
